ChatBot without ML
Basic Education counselling chatbot that uses predefined answers for user inputs. User inputs are first converted to lowercase and tokenized, followed by stemming. The stemmed words are then compared and mapped to an intent. Each intent consists of a list of responses, where one will be chosen using the "random" function in python.